+++ src/main/java/com/example/rebrain/dto/CardsSetsGetDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CardsSetsGetDto {
-    private Integer setId;
+    private Long setId;
     private String setTitle;
     private String setDescription;
     private List<CardDto> cards;
