+++ src/main/java/com/example/rebrain/repositories/NoteRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface NoteRepo extends JpaRepository<NoteEntity, Integer> {
+public interface NoteRepo extends JpaRepository<NoteEntity, Long> {
     List<NoteEntity> findByTitle(String title);
 }
