+++ src/main/java/com/example/rebrain/repositories/TestRepo.java
@@ -3,5 +3,5 @@ package com.example.rebrain.repositories;
 import com.example.rebrain.entity.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestRepo extends JpaRepository<TestEntity, Integer> {
+public interface TestRepo extends JpaRepository<TestEntity, Long> {
 }
